Makefile: Move `xcodebuild` steps into Make
I needed to remove the `-isysroot` compiler flag in order to get the
static library to build in my Rust project without the

    ld: framework not found ColorSync for architecture x86_64

linker error (which is apparently caused by the fact that I'm building
against a 10.13 SDK on a machine running 10.12, since ColorSync exists
only on 10.13).

Not sure if there's a way to have Xcode not use the `-isysroot` flag and
let us continue to build using `xcodebuild`.

Just copy-pasted the build output including exported environment
variables. I removed machine-specific things like absolute paths to my
home directory and the hash after the DerivedData directory.

Not very clean, but it seems to work.